rustc_target: move in type definitions from ty::layout.
